requirements.txt
google-cloud-aiplatform
pillow
